order stuff

+++ api/src/main/java/com/agerday/order/api/item/ItemMapper.java
@@ -9,7 +9,7 @@ public class ItemMapper {
         return new Item(
                 createItemDto.getName(),
                 createItemDto.getDescription(),
-                (int) createItemDto.getPrice(),
+                createItemDto.getPrice(),
                 createItemDto.getAmount()
 
         );
